consolidate gamepad configs (DayKev suggestion)
