+++ src/ui/settings/gamepad-settings-ui-handler.ts
@@ -5,9 +5,7 @@ import { Device } from "#enums/device";
 import { SettingGamepad } from "#enums/setting-gamepad";
 import { TextStyle } from "#enums/text-style";
 import { UiMode } from "#enums/ui-mode";
-import { pad_dualshock } from "#inputs/pad-dualshock";
-import { pad_unlicensedSNES } from "#inputs/pad-unlicensed-snes";
-import { pad_xbox360 } from "#inputs/pad-xbox360";
+import { pad_dualshock, pad_unlicensedSNES, pad_xbox360 } from "#inputs/gamepad-configs";
 import {
   setSettingGamepad,
   settingGamepadBlackList,
